pyproject.toml
[project]
name = "codeintp"
version = "1.0.0-beta1"
description = "2kbit Code: INTP"
authors = [
    { name = "Abjust", email = "2533898108@qq.com" },
]
dependencies = [
    "requests>=2.31.0",
    "py-cpuinfo>=9.0.0",
    "psutil>=5.9.5",
    "ruamel.yaml",
    "pymysql",
    "pytz",
    "opencc",
    "nonebot2[fastapi]",
    "nonebot-adapter-console>=0.6.0",
    "nonebot-plugin-alconna>=0.37.1",
    "nonebot-adapter-onebot>=2.4.4",
    "nonebot-plugin-apscheduler>=0.5.0",
    "cryptography>=42.0.8",
]
requires-python = ">=3.10, <3.13"
readme = "README.md"
license = { text = "AGPL-3.0" }

